Restore CLI access
