+++ src/rarfile/__init__.py
@@ -3093,33 +3093,3 @@ def tool_setup(unrar=True, unar=True, bsdtar=True, sevenzip=True, force=False):
         raise RarCannotExec("Cannot find working tool")
     return CURRENT_SETUP
 
-
-def main(args):
-    """Minimal command-line interface for rarfile module.
-    """
-    import argparse
-    p = argparse.ArgumentParser(description=main.__doc__)
-    g = p.add_mutually_exclusive_group(required=True)
-    g.add_argument("-l", "--list", metavar="<rarfile>",
-                   help="Show archive listing")
-    g.add_argument("-e", "--extract", nargs=2,
-                   metavar=("<rarfile>", "<output_dir>"),
-                   help="Extract archive into target dir")
-    g.add_argument("-t", "--test", metavar="<rarfile>",
-                   help="Test if a archive is valid")
-    cmd = p.parse_args(args)
-
-    if cmd.list:
-        with RarFile(cmd.list) as rf:
-            rf.printdir()
-    elif cmd.test:
-        with RarFile(cmd.test) as rf:
-            rf.testrar()
-    elif cmd.extract:
-        with RarFile(cmd.extract[0]) as rf:
-            rf.extractall(cmd.extract[1])
-
-
-if __name__ == "__main__":
-    main(sys.argv[1:])
-
