+++ src/rarfile/__main__.py
@@ -1,0 +1,7 @@
+"""Launch CLI"""
+
+import sys
+
+from .cli import main
+
+main(sys.argv[1:])
